ASC-1146 Update Dependecy Requirements
Updated the requirements to utilize the latest release of "rpc-zigzag".
I sorted the "requirements.txt" file alphabetically so that we don't
accidentally repeat requirements. (This happended before)

Also, I loosened the requirement on "rpc-zigzag" version a tad so at least
new patch versions are automatically consumed without always having to update
this repo. We still want to pin hard to the minor version to guarantee
compatibility with the JUniXML XSD from "rpc-zigzag"